First successful balance: Mega port, voltage-mode control, FG filter
- Switch controller to Arduino Mega 2560 (5V logic drives the BLDC-3640
  DIR/BRAKE pull-ups directly); keep ESP32 sketches in firmware/legacy_esp32
- Measured on hardware: motor polarity, IMU axes (atan2(ay,az), gyro X),
  gear ratio 6, control sign +1 (sign_test), open-loop step response
  (step_test: voltage-mode driver, tau ~0.18 s with bare wheel)
- Voltage/torque command: duty = (omega_w + tau*a)/omega_fs instead of an
  integrated speed setpoint (sim 0/20 -> 20/20 at tau 0.18)
- Reject FG glitch edges (<700 us) that read as 30k+ rpm around DIR flips
- Hardware-tuned Kd 500: balanced for minutes, tilt SD 0.34 deg
- Add tools/imu_viewer.html (Web Serial axis finder)

+++ firmware/reaction_wheel_balance/complementary_filter.h
@@ -5,7 +5,7 @@
 // timescales) with a gyro rate integration (smooth short-term, drifts long
 // term) into a single tilt angle estimate.
 // =============================================================================
-#include <cmath>
+#include <math.h>
 
 class ComplementaryFilter {
 public:
@@ -37,5 +37,5 @@ private:
 // tilt plane, via atan2(num, den). When upright, den should read close to
 // +1 g and num close to 0 g.
 inline float accelTiltAngle(float num, float den) {
-  return std::atan2(num, den);
+  return atan2f(num, den);
 }
